+++ packages/cli/src/list/data-set-members/DataSetMembers.definition.ts
@@ -32,6 +32,11 @@ export const ListDataSetMembersDefinition: ICommandDefinition = {
             required: true,
         },
     ],
+    options: [{
+        name: "pattern",
+        description: "A wildcard pattern used to filter the returned entries (supports '*' and '?')",
+        type: "string"
+    }],
     profile: { optional: ["ssh"] },
     outputFormatOptions: true,
 };